% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_recs.R
\name{taxon_children}
\alias{taxon_children}
\title{Count number of child taxa of a given rank for a patricular taxon.}
\usage{
taxon_children(taxa, child_rank = "species")
}
\arguments{
\item{child_rank}{character, taxonomic rank of child taxon to count (defaults
to species)}

\item{taxon_name}{Character, name of the taxon to gather information on.}
}
\value{
A numeric vector or length one that contains the number of child
taxa of the given rank
}
\description{
Count number of child taxa of a given rank for a patricular taxon.
}
\examples{
taxon_children(taxon_name="Plasmodium")
}
