% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_recs.R
\name{taxon_records}
\alias{taxon_records}
\title{Count number of records from a given taxon in an NCBI database.}
\usage{
taxon_records(taxa, db, additional_terms = NULL)
}
\arguments{
\item{db}{Character, NCBI database from which to gather data.}

\item{additional_terms}{Character, additional term.}

\item{taxon_name}{Character, name of the taxon to gather information on.}
}
\value{
A numeric vector or length one that contains the number of records
matching the criteria
}
\description{
Count number of records from a given taxon in an NCBI database.
}
\examples{
taxon_records(taxon_name="Plasmodium", db="genome")
}
