% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tidy_taxonomy}
\alias{tidy_taxonomy}
\title{Produce a tidydata summary of a taxomic group}
\usage{
tidy_taxonomy(subtree, lowest_rank, higher_ranks)
}
\arguments{
\item{subtree}{character, subtree from which to extract names}

\item{lowest_rank}{Character, name of the lowest taxonomic rank to be included in the data.frame}

\item{higher_ranks}{Character, taxonomic ranks between "taxon_rank" and subtree}
}
\value{
a data.frame containing a summary of the given taxonmic data in 'tidy'
format.
}
\description{
Produce a tidydata summary of a taxomic group
}
\examples{
ape_taxonomy <- taxon_df(subtree="apes", lowest_rank="species", higher_ranks= c("family", "genus"))
}
